Extract PointCloudSelectionHandler from PointCloudCommon
